+++ R/rspm.R
@@ -46,7 +46,7 @@ renv_rspm_transform_impl <- function(url) {
   mirrors <- getCRANmirrors(local.only = TRUE)
   urls <- mirrors$URL
 
-  # include RStudio URLs
+  # also ignore some RStudio URLs
   rstudio <- c(
     "http://cran.rstudio.com",
     "http://cran.rstudio.org",
@@ -122,22 +122,45 @@ renv_rspm_platform <- function() {
     )
 
     id <- properties$ID %||% ""
-    id_like <- properties$ID_LIKE %||% ""
-    version_codename <- properties$VERSION_CODENAME %||% ""
-    version_id <- properties$VERSION_ID %||% ""
 
-    if (id == "ubuntu")
-      return(version_codename)
-
-    if (id == "centos")
-      return(paste(id, version_id, sep = ""))
-
-    if (grepl("\\bsuse\\b", id_like)) {
-      parts <- strsplit(version_id, ".", fixed = TRUE)
-      return(paste("opensuse", parts[[1]], sep = ""))
-    }
+    case(
+      identical(id, "ubuntu") ~ renv_rspm_platform_ubuntu(properties),
+      identical(id, "centos") ~ renv_rspm_platform_centos(properties),
+      grepl("\\bsuse\\b", id) ~ renv_rspm_platform_suse(properties)
+    )
 
   }
+
+}
+
+renv_rspm_platform_ubuntu <- function(properties) {
+
+  codename <- properties$VERSION_CODENAME
+  if (is.null(codename))
+    return(NULL)
+
+  codename
+
+}
+
+renv_rspm_platform_centos <- function(properties) {
+
+  id <- properties$VERSION_ID
+  if (is.null(id))
+    return(NULL)
+
+  paste0("centos", id)
+
+}
+
+renv_rspm_platform_suse <- function(properties) {
+
+  id <- properties$VERSION_ID
+  if (is.null(id))
+    return(NULL)
+
+  parts <- strsplit(id, ".", fixed = TRUE)[[1L]]
+  paste0("opensuse", parts[[1L]])
 
 }